requirements.txt
python-telegram-bot==20.7
huggingface-hub==0.20.0
tavily-python==0.3.0
python-dotenv==1.0.0
aiohttp==3.9.1

